LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/duplicate-tag-rejects-capture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_true: ready not rejected before updateCallbackDone expected true got false

FAIL Two different elements with the same name in the old DOM should skip the transition promise_test: Unhandled rejection with value: undefined
